+++ source/graphs/single_source_shortest_paths/dag_shortest_paths.py
@@ -6,7 +6,7 @@ from source.graphs.single_source_shortest_paths.helpers.relax import relax
 
 
 def dag_shortest_paths(
-    g: Graph, w: dict[tuple[Vertex, Vertex], int], s: Vertex
+    g: Graph, w: dict[tuple[Vertex, Vertex], float], s: Vertex
 ) -> None:
     topologic = topological_sort(g)
     initialize_single_source(g, s)
