+++ source/graphs/single_source_shortest_paths/helpers/relax.py
@@ -1,7 +1,7 @@
 from source.datastructures.graph import Vertex
 
 
-def relax(u: Vertex, v: Vertex, w: dict[tuple[Vertex, Vertex], int]) -> None:
+def relax(u: Vertex, v: Vertex, w: dict[tuple[Vertex, Vertex], float]) -> None:
     if v.d > (distance := u.d + w[(u, v)]):
         v.d = distance
         v.pi = u
